@@ -16,7 +16,7 @@ const route = Router(); /* eslint new-cap: 0 */
 /*
  This file include all verdaccio only API(Web UI), for npm API please see ../endpoint/
 */
-module.exports = function(config: Config, auth: IAuth, storage: IStorageHandler) {
+export default function(config: Config, auth: IAuth, storage: IStorageHandler) {
   Search.configureStorage(storage);
 
   // validate all of these params as a package name
@@ -43,4 +43,4 @@ module.exports = function(config: Config, auth: IAuth, storage: IStorageHandler)
   // We will/may replace current token with JWT in next major release, and it will not expire at all(configurable).
 
   return route;
-};
+}
